feat: 🎸 게시글 삭제 - 관계 맺은 운동기록, 참여, 평가에 대해서도 삭제

+++ backend/src/main/java/com/ssafy/woonana/domain/service/board/BoardService.java
@@ -9,29 +9,37 @@ import com.ssafy.woonana.domain.model.entity.participation.Participation;
 import com.ssafy.woonana.domain.model.entity.user.User;
 import com.ssafy.woonana.domain.repository.board.BoardRepository;
 import com.ssafy.woonana.domain.repository.board.ExerciseRepository;
+import com.ssafy.woonana.domain.repository.evaluation.EvaluationRepository;
+import com.ssafy.woonana.domain.repository.exercise.ExerciseLogRepository;
 import com.ssafy.woonana.domain.repository.participation.ParticipationRepository;
 import com.ssafy.woonana.domain.repository.user.UserRepository;
 import com.ssafy.woonana.error.exception.custom.ParticipationDuplicateException;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
 @Service
+@Slf4j
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class BoardService {
-
+    @PersistenceContext EntityManager em;
     private final BoardRepository boardRepository;
     private final ExerciseRepository exerciseRepository;
+    private final ExerciseLogRepository exerciseLogRepository;
     private final UserRepository userRepository;
     private final ParticipationRepository participationRepository;
+    private final EvaluationRepository evaluationRepository;
     private final AwsS3Service awsS3Service;
 
     @Transactional
@@ -105,10 +113,20 @@ public class BoardService {
         return new BoardDetailResponse(findBoard, findBoardUser);
     }
     @Transactional
-    public void deleteBoard(Long boardId) {
-        // TODO: 참여하기로한 멤버가 있는 경우
-        // TODO: 완료하지 않은 운동인 경우 운동 기록, 참여 등에도 삭제가 반영 되는지 확인 필요
-        boardRepository.deleteById(boardId);
+    public void deleteBoard(Long boardId, Long userId) {
+        Board findBoard = boardRepository.findById(boardId).get();
+        if (findBoard.getUser().getUserId() == userId) {
+            // TODO: 참여하기로한 멤버가 있는 경우
+            // TODO: 완료하지 않은 운동인 경우 운동 기록, 참여 등에도 삭제가 반영 되는지 확인 필요
+            exerciseLogRepository.deleteLogByBoardId(boardId);
+
+            participationRepository.deleteParticipationByBoardId(boardId);
+
+            evaluationRepository.deleteEvaluationByBoardId(boardId);
+
+            boardRepository.deleteById(boardId);
+        }
+
     }
 
     private void isTimeOverlapped(Long userId, LocalDateTime meetStartDate, LocalDateTime meetEndDate) {
